Rename nulldatatable.txt to index.tsx and change imports
